+++ test/client.ts
@@ -1445,5 +1445,88 @@ describe('client', () => {
       assert(fragmentDefinitionsMap.hasOwnProperty('authorDetails'));
       assert.equal(fragmentDefinitionsMap['authorDetails'].length, 1);
     });
+
+    it('should not mutate the input document when querying', () => {
+      const client = new ApolloClient();
+
+      const fragments = createFragment(gql`
+        fragment authorDetails on Author {
+          author {
+            firstName
+            lastName
+          }
+        }`);
+      const query = gql`{ author { ...authorDetails } }`;
+      const initialDefinitions = query.definitions;
+      client.query({query, fragments});
+      assert.equal(query.definitions, initialDefinitions);
+    });
+  });
+
+  it('should pass a network error correctly on a mutation', (done) => {
+    const mutation = gql`
+      mutation {
+        person {
+          firstName
+          lastName
+        }
+      }`;
+    const data = {
+      person: {
+        firstName: 'John',
+        lastName: 'Smith',
+      },
+    };
+    const networkError = new Error('Some kind of network error.');
+    const client = new ApolloClient({
+      networkInterface: mockNetworkInterface({
+        request: { query: mutation },
+        result: { data },
+        error: networkError,
+      }),
+    });
+
+    client.mutate({ mutation }).then((result) => {
+      done(new Error('Returned a result when it should not have.'));
+    }).catch((error) => {
+      const apolloError = error as ApolloError;
+      assert(apolloError.networkError);
+      assert.equal(apolloError.networkError.message, networkError.message);
+      done();
+    });
+  });
+
+  it('should pass a GraphQL error correctly on a mutation', (done) => {
+    const mutation = gql`
+      mutation {
+        newPerson {
+          person {
+            firstName
+            lastName
+          }
+        }
+      }`;
+    const data = {
+      person: {
+        firstName: 'John',
+        lastName: 'Smith',
+      },
+    };
+    const errors = [ new Error('Some kind of GraphQL error.') ];
+    const client = new ApolloClient({
+      networkInterface: mockNetworkInterface({
+        request: { query: mutation },
+        result: { data, errors },
+      }),
+    });
+    client.mutate({ mutation }).then((result) => {
+      done(new Error('Returned a result when it should not have.'));
+    }).catch((error) => {
+      const apolloError = error as ApolloError;
+      assert(apolloError.graphQLErrors);
+      assert.equal(apolloError.graphQLErrors.length, 1);
+      assert.equal(apolloError.graphQLErrors[0].message, errors[0].message);
+      done();
+    });
   });
 });